+++ KMSCalendar/KMSCalendar/Services/Data/IDataStore.cs
@@ -3,7 +3,7 @@ using System.Threading.Tasks;
 
 using KMSCalendar.Models.Data;
 
-namespace KMSCalendar.Services
+namespace KMSCalendar.Services.Data
 {
     public interface IDataStore<T> where T : TableData
     {
