+++ KMSCalendar/KMSCalendar/Services/Data/MockDataStore.cs
@@ -4,7 +4,7 @@ using System.Threading.Tasks;
 
 using KMSCalendar.Models.Data;
 
-namespace KMSCalendar.Services
+namespace KMSCalendar.Services.Data
 {
     public class MockDataStore<T> : IDataStore<T> where T : TableData
     {
