+++ KMSCalendar/KMSCalendar/ViewModels/AssignmentDetailViewModel.cs
@@ -4,7 +4,7 @@ using System.Windows.Input;
 using Xamarin.Forms;
 
 using KMSCalendar.Models.Data;
-using KMSCalendar.Services;
+using KMSCalendar.Services.Data;
 
 namespace KMSCalendar.ViewModels
 {
