+++ KMSCalendar/KMSCalendar/ViewModels/ClassSearchViewModel.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using Xamarin.Forms;
 
 using KMSCalendar.Models.Data;
-using KMSCalendar.Services;
+using KMSCalendar.Services.Data;
 
 namespace KMSCalendar.ViewModels
 {
